+++ src/Componentes/CardVideo/CardVideo.jsx
@@ -1,0 +1,24 @@
+import React from "react";
+import InfosUsuario from "../InfosUsuario/InfosUsuario";
+import "./styles.css"
+
+const CardVideo = () => {
+    const titulo = "Título do vídeo";
+    
+    function reproduzVideo() {
+      alert("O vídeo está sendo reproduzido");
+    }
+ 
+      return (
+    <div className="box-pagina-principal" onClick={reproduzVideo}>
+    <img src="https://picsum.photos/400/400?a=1 " alt="" />
+    <h4>{titulo}</h4>
+   <InfosUsuario/>
+  
+  </div>
+  )
+}
+
+
+
+export default CardVideo;
