+++ src/Componentes/InfosUsuario/InfosUsuario.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import "./styles.css"
+
+const InfosUsuario = () => {
+  return <div>
+<img src="" alt="" />
+<h1>Mario</h1>
+  </div>;
+};
+
+export default InfosUsuario;
