+++ src/Componentes/InfosUsuario/styles.css
@@ -1,0 +1,5 @@
+*{
+    margin: 0;
+    padding: 0;
+}
+
